requirements.txt
openai
langchain
langchain-community
chromadb
tiktoken
python-dotenv
streamlit